Bug 1309834. r=wchen.
MozReview-Commit-ID: FO7bXlx6lv

UltraBlame original commit: 8704571a427f466c9143f46663862fb979429d7a

+++ parser/html/nsHtml5TreeBuilderHSupplement.h
@@ -138,7 +138,7 @@
 
 
 
-    bool EnsureBufferSpace(size_t aLength);
+    bool EnsureBufferSpace(int32_t aLength);
 
     void EnableViewSource(nsHtml5Highlighter* aHighlighter);
